+++ TFG/Model/Service/Tarifa/IServiceTarifa.cs
@@ -15,9 +15,9 @@ namespace Es.Udc.DotNet.TFG.Model.Service.Tarifas
         [Transactional]
         List<TarifaDTO> verTarifasDelDia(DateTime fecha);
         [Transactional]
-        long BuscarMejorTarifa(DateTime fecha);
+        double BuscarMejorTarifa(DateTime fecha);
         [Transactional]
-        long BuscarpeorTarifa(DateTime fecha);
+        double BuscarpeorTarifa(DateTime fecha);
         [Transactional]
         List<TarifaDTO> OrdenarMejorPrecioTarifasDelDia(DateTime fecha);
         [Transactional]
